# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Recall API
  version: "1"
  description: |
    HTTP REST + SSE surface for the Recall Overwatch metrics app.

    Available in two flavors:
      * **Server binary** (`Recall-server`) — exposes this API by default
      * **Wails desktop binary with `--server`** — same API, same routes

    Conventions:
      * Every JSON endpoint is mounted under the `/api/v1/` prefix —
        the version is bumped on breaking changes.
      * Resources are nouns; sub-resources hang off the parent
        (`/matches/{match_key}/visibility`, `/settings/tesseract`).
      * Methods reflect intent: `GET` to read, `PUT` to replace/upsert,
        `DELETE` to clear or wipe, `POST` to start a server-side action
        that doesn't map to a single resource (the parse run).
      * Writes return `204 No Content` when there is no useful body
        to echo back; `202 Accepted` for actions whose effect is
        asynchronous (parse).
      * Errors come back as `application/problem+json` (RFC 9457) with
        the appropriate 4xx/5xx status. `type` is a stable URI under
        `https://github.com/sound-barrier/recall/problems/{slug}` that
        clients can switch on; body-validation problems carry a
        per-field `errors` array, and the data-update 422 carries a
        `failed_assets` list (RFC 9457 §3.2 extension members). The one
        exception is `405 Method Not Allowed`, emitted plain by the
        router.
...

...
    `RECALL_SERVER_ADDR`.
  license:
...
servers:
...
paths:
...
  /api/v1/parses:
    post:
      operationId: ParseScreenshots
      tags: [Pipeline]
      summary: Run a one-shot parse
      description: |
        Walks the configured screenshots folder, OCRs anything not
        already in the DB, and writes the merged rows. Idempotent —
        re-running won't duplicate matches already on record.

        `?scope=all` switches into "re-parse all": every PNG in the
        watched folder gets re-OCR'd regardless of whether it's
        already in the per-type tables. Use after a parser-tightening
        release (e.g. the hero-fuzzy-match length-gate that stopped
        Miyazaki being attributed to Mei) to retroactively correct
        older rows. The Upsert clauses are idempotent on filename so
        existing rows are rewritten in place; match annotations /
        queue / play-mode overrides / hidden flags / reviews all key
        on match_key and survive the re-parse. ~1 s per screenshot.

        Runs the parse in a BACKGROUND goroutine and returns
        `202 Accepted` **immediately** — the request is not held open
        for the OCR loop. Subscribe to `GET /api/v1/events` for progress
        + completion (`parse-progress` / `parse-complete`), and poll
        `GET /api/v1/parses/active` to resync after a reconnect or page
        reload (SSE is not replayed). Re-fetch `GET /api/v1/matches` on
        completion.

        `409 Conflict` when a parse is already in flight (single-flight,
        fail-fast) or the screenshots folder is unset/unreadable.
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [new, all]
          description: |
            `new` (default) parses only files not in the DB.
            `all` re-OCRs every file. Other values return 400.
      responses:
        "202":
          description: Parse accepted and started; subscribe to events for progress + completion.
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/parses/active:
    get:
      operationId: GetActiveParse
      tags: [Pipeline]
      summary: Active-parse status snapshot
      description: |
        The resync anchor for the async parse pipeline. SSE messages are
        not replayed on connect, so a client that reconnects or reloads
        mid-parse reads this to restore "is a parse running, and how far
        along." `running=false` when idle.
      responses:
        "200":
          description: Current parse run-state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveParse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: CancelParse
      tags: [Pipeline]
      summary: Stop an in-flight parse
      description: |
        Cancels the parse run currently in flight. The OCR loop reads
        the cancellation flag between screenshots, so the file
        currently in tesseract finishes before the loop unwinds —
        cancellation lands at the next between-files boundary, not
        synchronously with this response.

        Any per-screenshot rows already committed to SQLite stay put;
        a follow-up `GET /api/v1/matches` reflects the partial state.
        The frontend learns about completion via the SSE
        `parse-cancelled` event so it can flip the Stop button back
        to Run without polling.

        `202` on success, `409` when no parse is in flight.
      responses:
        "202":
          description: Cancellation signalled; SSE `parse-cancelled` follows once the loop unwinds.
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/screenshots/pending-count:
    get:
      operationId: GetPendingScreenshotCount
      tags: [Pipeline]
      summary: Count of screenshots awaiting parse
      description: |
        Returns the number of files in the configured folder that are
        not yet present in any per-screenshot-type table's `filename`.
      responses:
        "200":
          description: Count.
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
                    minimum: 0
                    example: 3
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status]
      description: |
        RFC 9457 (Problem Details for HTTP APIs) error body, returned as
        `application/problem+json` for every 4xx/5xx the API surfaces. `type`
        is a stable URI under
        `https://github.com/sound-barrier/recall/problems/` that clients can
        switch on; `instance` is the request path. Members beyond those listed
        here are RFC 9457 §3.2 extension members (the schema stays open).
      properties:
        type:
          type: string
          format: uri
          description: A URI identifying the problem class.
          example: https://github.com/sound-barrier/recall/problems/invalid-body
        title:
          type: string
          description: Short, human-readable summary of the problem class.
          example: Bad Request
        status:
          type: integer
          description: The HTTP status code, duplicated in the body for convenience.
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: The request path that produced the problem.
        errors:
          type: array
          description: |
            §3.2 extension — per-field validation failures, present on a
            malformed request body.
          items:
            $ref: "#/components/schemas/FieldError"
        failed_assets:
          type: array
          description: |
            §3.2 extension — the asset(s) whose SHA-256 sidecar verification
            failed (the data-update 422).
          items:
            type: string
    FieldError:
      type: object
      required: [field, detail]
      description: One offending request field and why it was rejected.
      properties:
        field:
          type: string
          description: The request-body field name.
        detail:
          type: string
          description: Why the field was rejected.
    ActiveParse:
      type: object
      required: [running, done, total, scope]
      description: |
        Snapshot of the in-flight parse run. The resync anchor for the
        async parse pipeline — read after an SSE reconnect or a page
        reload to restore the parse panel without replaying events.
      properties:
        running:
          type: boolean
          description: Whether a parse is currently in flight.
        done:
          type: integer
          minimum: 0
          description: Screenshot files OCR'd so far this run.
        total:
          type: integer
          minimum: 0
          description: Total files to process this run (0 when idle).
        scope:
          type: string
          description: '"new" (default parse) or "all" (re-parse); empty when idle.'
...
  responses:
    BadRequest:
      description: Malformed request body or query parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    Conflict:
      description: |
        The request was syntactically valid, but the resource state or
        a payload value prevents the action (e.g. duplicate profile
        name, screenshots directory not configured, invalid Tesseract
        binary path, non-candidate resolution target).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InternalError:
      description: Unhandled server-side error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...